+++ packages/utils/src/spm/index.ts
@@ -1,0 +1,11 @@
+/*
+ * @Author: @memo28.repo
+ * @Date: 2024-08-30 17:18:15
+ * @LastEditTime: 2024-08-30 17:44:37
+ * @Description: 解析分享参数规范 spm
+ * @FilePath: /memoRepo/packages/utils/src/spm/index.ts
+ */
+
+export { Spm } from './impl'
+export type { SpmImpl, SpmItem } from './interface'
+